requirements.txt
nltk
jiwer
rouge-score
bert-score
unbabel-comet
openpyxl